+++ src/client/route/admin/server/ping_format/mod.rs
@@ -42,7 +42,7 @@ pub fn ServerAdminPingFormat(guild_id: u64) -> Element {
             }
 
             // Only run resource if we need to fetch
-            let needs_fetch = guild.read().as_ref().map(|g| g.guild_id as u64) != Some(guild_id);
+            let needs_fetch = guild.read().as_ref().map(|g| g.guild_id) != Some(guild_id);
 
             if needs_fetch {
                 should_fetch.set(true);
